move reranking functions into Reranker class
