+++ src/rank_llm/rerank/__init__.py
@@ -3,6 +3,7 @@ import logging
 from .api_keys import get_azure_openai_args, get_openai_api_key
 from .identity_reranker import IdentityReranker
 from .rankllm import PromptMode, RankLLM
+from .reranker import Reranker
 
 logging.basicConfig(
     level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
@@ -14,4 +15,5 @@ __all__ = [
     "get_azure_openai_args",
     "get_openai_api_key",
     "PromptMode",
+    "Reranker",
 ]
