+++ src/rank_llm/rerank/listwise/rank_gpt.py
@@ -8,6 +8,7 @@ from tqdm import tqdm
 
 from rank_llm.data import Request, Result
 from rank_llm.rerank import PromptMode
+
 from .listwise_rankllm import ListwiseRankLLM
