+++ src/rank_llm/rerank/listwise/rank_listwise_os_llm.py
@@ -113,11 +113,11 @@ class RankListwiseOSLLM(ListwiseRankLLM):
         logging: bool = False,
         **kwargs: Any,
     ) -> List[Result]:
-        top_k_retrieve: int = kwargs.get('top_k_retrieve', 50)
-        window_size: int = kwargs.get('window_size', 20)
+        top_k_retrieve: int = kwargs.get("top_k_retrieve", 50)
+        window_size: int = kwargs.get("window_size", 20)
         window_size = min(window_size, top_k_retrieve)
-        step: int = kwargs.get('step', 10)
-        populate_exec_summary: bool = kwargs.get('populate_exec_summary', False)
+        step: int = kwargs.get("step", 10)
+        populate_exec_summary: bool = kwargs.get("populate_exec_summary", False)
 
         if self._vllm_batched:
             # reranking using vllm
