+++ src/rank_llm/retrieve/pyserini_retriever.py
@@ -22,7 +22,6 @@ from pyserini.search import (
 from tqdm import tqdm
 
 from rank_llm.data import Candidate, DataWriter, Query, Request
-
 from rank_llm.retrieve.indices_dict import INDICES
 from rank_llm.retrieve.topics_dict import TOPICS
